+++ src/pc/1.14/ChunkColumn.js
@@ -1,7 +1,7 @@
 const SmartBuffer = require('smart-buffer').SmartBuffer
-const ChunkSection = require('./ChunkSection')
-const constants = require('../common/constants')
 const BitArray = require('../common/BitArray')
+const ChunkSection = require('../common/CommonChunkSection')(BitArray)
+const constants = require('../common/constants')
 const varInt = require('../common/varInt')
 
 // wrap with func to provide version specific Block
